+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - previously deprecated `Worker.Start()` and `WorkerPool.Start()` removed in favour of `Worker.Run()`
   and `WorkerPool.Run()` - please check documentation as they are slightly different
 - `Job.Priority` changed its type from `int16` to `JobPriority` that is a wrapper type for `int16`
+- `gue/adapter/exponential.Default` became `gue.DefaultExponentialBackoff`
+- `gue/adapter/exponential.New()` became `gue.NewExponentialBackoff()`
 
 ### New
 
@@ -22,3 +24,5 @@
   - `JobPriorityLow`
   - `JobPriorityLowest`
 - `WorkerPool.WorkOne` method, can be useful for testing purpose mostly
+- backoff implementation may return negative value to discard errored job immediately
+  - `gue.BackoffNever` backoff implementation discards the job on the first error